Seven Segment, Rotary Encoder, I2C
This example code is written for a 4-PCB setup. The 4 displays/rotary
encoders are used as user interface for an i2c connected master that
controls lamps. The first number is a 6-bit-address, the second number
is a dim value for values from 0 to 100 ("A0" due to the limitation of
only having two digits).